+++ lib/ts/recipe/thirdpartyemailpassword/index.ts
@@ -103,6 +103,10 @@ export default class Wrapper {
         return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.getUserCount();
     }
 
+    static updateEmailOrPassword(input: { userId: string; email?: string; password?: string }) {
+        return Recipe.getInstanceOrThrowError().recipeInterfaceImpl.updateEmailOrPassword(input);
+    }
+
     static createEmailVerificationToken(userId: string) {
         return Recipe.getInstanceOrThrowError().createEmailVerificationToken(userId);
     }
@@ -169,6 +173,8 @@ export let getUsersNewestFirst = Wrapper.getUsersNewestFirst;
  * @deprecated Use supertokens.getUserCount(...) function instead IF using core version >= 3.5
  *   */
 export let getUserCount = Wrapper.getUserCount;
+
+export let updateEmailOrPassword = Wrapper.updateEmailOrPassword;
 
 export let Google = Wrapper.Google;